+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Added
+
+- Support for escape and unicode characters [#98](https://github.com/ufirstgroup/ymlr/pull/98)
+
 <!--------------------- Don't add new entries after this line --------------------->
 
 ### Added